js_fix_cleaning: refactored names to kebab-case
