+++ src/app/model/vehicle.ts
