+++ src/app/service/student.service.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Student} from "../model/Student";
+import {Student} from "../model/student";
 import {HttpClient} from "@angular/common/http";
 import {CrudService} from "./crud-service";
 import {Constraints} from "./constraints";
